update: delete-piu-service

+++ src/modules/pius/services/DeletePiuService.ts
@@ -7,6 +7,7 @@ import Piu from '../infra/typeorm/entities/Piu';
 
 interface Request{
   piu_id: string
+  user_id: string
 }
 
 @injectable()
@@ -16,13 +17,17 @@ export default class DeletePiuService {
     private piusRepository: IPiusRepository,
   ) {}
 
-  public async execute({ piu_id }: Request): Promise<void> {
+  public async execute({ piu_id, user_id }: Request): Promise<void> {
     const piu = await this.piusRepository.findById(piu_id)
 
     if (!piu){
       throw new AppError('Piu not found', 404)
     } else {
-      await this.piusRepository.delete(piu_id);
+      if ( piu.user_id != user_id){
+        throw new AppError('User logged does not own Piu', 401)
+      } else {
+        await this.piusRepository.delete(piu_id);
+      }
     }
   }
 }
